# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/invite": {
      "post": {
        "operationId": "InviteController_createInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Invite"
        ]
      },
      "get": {
        "operationId": "InviteController_getInvites",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InviteResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Invite"
        ]
      }
    },
    "/api/invite/public/{code}": {
      "get": {
        "operationId": "InviteController_getPublicInvite",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Invite"
        ]
      }
    },
    "/api/invite/{code}": {
      "get": {
        "operationId": "InviteController_getInvite",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Invite"
        ]
      },
      "delete": {
        "operationId": "InviteController_deleteInvite",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invite"
        ]
      }
    },
...
    "/api/file-upload": {
      "post": {
        "operationId": "FileUploadController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "FileUpload"
        ]
      }
    },
    "/api/file-upload/{id}": {
      "delete": {
        "operationId": "FileUploadController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "FileUpload"
        ]
      }
    },
...
    "/api/friends/status/{userId}": {
      "get": {
        "operationId": "FriendsController_getFriendshipStatus",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipStatusDto"
                }
              }
            }
          }
        },
        "summary": "Get friendship status with a specific user",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/request/{userId}": {
      "post": {
        "operationId": "FriendsController_sendFriendRequest",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipDto"
                }
              }
            }
          }
        },
        "summary": "Send a friend request to a user",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/accept/{id}": {
      "post": {
        "operationId": "FriendsController_acceptFriendRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipDto"
                }
              }
            }
          }
        },
        "summary": "Accept a friend request",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/decline/{id}": {
      "delete": {
        "operationId": "FriendsController_declineFriendRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Decline a friend request",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/cancel/{id}": {
      "delete": {
        "operationId": "FriendsController_cancelFriendRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Cancel a sent friend request",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/{id}": {
      "delete": {
        "operationId": "FriendsController_removeFriend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Remove a friend (unfriend)",
        "tags": [
          "Friends"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kraken API",
    "description": "Kraken Chat Application API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "maxUses": {
            "type": "number",
            "minimum": 1
          },
          "validUntil": {
            "format": "date-time",
            "type": "string"
          },
          "communityIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "communityIds"
        ]
      },
      "InviteCreatorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "displayName"
        ]
      },
      "InviteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "defaultCommunityId": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxUses": {
            "type": "number",
            "nullable": true
          },
          "uses": {
            "type": "number"
          },
          "validUntil": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "usedByIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "disabled": {
            "type": "boolean"
          },
          "createdBy": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/InviteCreatorDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "code",
          "createdById",
          "defaultCommunityId",
          "maxUses",
          "uses",
          "validUntil",
          "createdAt",
          "usedByIds",
          "disabled"
        ]
      },
...
      "CreateFileUploadDto": {
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string",
            "enum": [
              "USER_AVATAR",
              "USER_BANNER",
              "COMMUNITY_BANNER",
              "COMMUNITY_AVATAR",
              "MESSAGE_ATTACHMENT",
              "CUSTOM_EMOJI",
              "REPLAY_CLIP"
            ]
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "resourceType"
        ]
      },
      "FileUploadResponseDto": {
        "type": "object",
        "properties": {
          "fileType": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO",
              "AUDIO",
              "DOCUMENT",
              "OTHER"
            ]
          },
          "resourceType": {
            "type": "string",
            "enum": [
              "USER_AVATAR",
              "USER_BANNER",
              "COMMUNITY_BANNER",
              "COMMUNITY_AVATAR",
              "MESSAGE_ATTACHMENT",
              "CUSTOM_EMOJI",
              "REPLAY_CLIP"
            ]
          },
          "storageType": {
            "type": "string",
            "enum": [
              "LOCAL",
              "S3",
              "AZURE_BLOB"
            ]
          },
          "id": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "checksum": {
            "type": "string"
          },
          "uploadedById": {
            "type": "string",
            "nullable": true
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "storagePath": {
            "type": "string"
          }
        },
        "required": [
          "fileType",
          "resourceType",
          "storageType",
          "id",
          "filename",
          "mimeType",
          "size",
          "checksum",
          "uploadedById",
          "uploadedAt",
          "deletedAt",
          "resourceId",
          "storagePath"
        ]
      },
...
      "FriendshipStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true,
            "enum": [
              "PENDING",
              "ACCEPTED",
              "BLOCKED"
            ]
          },
          "direction": {
            "type": "string",
            "enum": [
              "sent",
              "received"
            ],
            "nullable": true
          },
          "friendshipId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "status",
          "direction",
          "friendshipId"
        ]
      },
      "FriendshipDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "BLOCKED"
            ]
          },
          "id": {
            "type": "string"
          },
          "userAId": {
            "type": "string"
          },
          "userBId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "status",
          "id",
          "userAId",
          "userBId",
          "createdAt"
        ]
      },
...
    }
  }
}